renamed error_audit_mgmt.py to auditerror_mgmt.py for easier reading and more aligned to function. Moved out lists_generated_data.py from common as it is not longer needed. Renamed all existing references to auditerror_mgmt.py to align with update. Updated Platform-Usage-Basics.md to better explain platform setup and usage
