+++ datatier_actions/platform.py
@@ -8,8 +8,8 @@ from common.platform_modules import load_platform_capabilities
 import common.platform_settings
 from common.platform_settings import build_platform_variables
 from common.platform_settings import build_platform_config
-import common.error_audit_mgmt
-from common.error_audit_mgmt import process_auditerror_details
+import common.auditerror_mgmt
+from common.auditerror_mgmt import process_auditerror_details
 
 def query_platformdata_general(platform_vars, platform_settings, sql_connection, table_name)->list:
     try:
